Normalize color temperature to 0-1 range instead of Kelvin
Changed getColorTemperature() to return normalized values (0=warm, 1=cold)
instead of Kelvin values, matching the input format expected by
setColorTemperature(). This ensures consistency between reading and
writing color temperature values using HomeKit's mired units with
device-specific min/max ranges.

- HMCharacteristic: Normalize mired to 0-1 using metadata min/max
- EntityStorageItem: Change colorTemperature type from Int to Float
- EntityCharacteristicView: Display as percentage (0-100%)

+++ Sources/HAModels/EntityStorageItem.swift
@@ -31,9 +31,9 @@ public struct EntityStorageItem: Equatable, Sendable, Codable, CustomStringConve
     /// Light brightness level (0-100)
     public let brightness: Int?
 
-    /// Color temperature in Kelvin (K)
-    /// Converted from HomeKit's mired value: K = 1,000,000 / mired
-    public let colorTemperature: Int?
+    /// Color temperature normalized to 0...1 (warm...cold)
+    /// Normalized from HomeKit's mired value using device min/max capabilities
+    public let colorTemperature: Float?
 
     /// RGB color value
     public let color: RGB?
@@ -90,7 +90,7 @@ public struct EntityStorageItem: Equatable, Sendable, Codable, CustomStringConve
         illuminance: Measurement<UnitIlluminance>? = nil,
         isDeviceOn: Bool? = nil,
         brightness: Int? = nil,
-        colorTemperature: Int? = nil,
+        colorTemperature: Float? = nil,
         color: RGB? = nil,
         isContactOpen: Bool? = nil,
         isDoorLocked: Bool? = nil,
